translate_image/requirements.txt
asgiref==3.4.1
backports.functools-lru-cache==1.6.4
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
Cython==3.0.0a8
fastapi==0.66.0
future==0.18.2
google-api-core==1.30.0
google-auth==1.32.1
google-cloud-core==1.7.1
google-cloud-translate==3.2.1
google-transliteration-api==1.0.3
googleapis-common-protos==1.53.0
grpcio==1.38.1
h11==0.12.0
idna==2.10
importlib-metadata==4.6.1
numpy==1.21.0
packaging==21.0
pbr==5.6.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
pytz==2021.1
regex==2020.10.11
requests==2.25.1
rsa==4.7.2
scipy==1.7.0
selenium==3.141.0
six==1.16.0
starlette==0.14.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
zipp==3.5.0
